Rewrite as a WebExtension
This is a major change that reworks the build architecture to emit a
WebExtension that can be used by Chrome, Firefox and Opera instead of
having a separate build system for each browser. For easier maintenance
this will be the only extension that I will maintain going forward. This
means:

- The Safari extension is no longer supported. Note that the recent
Apple Extension Store changes already forced this.
- The bookmarklet is also no longer supported.
- Hence, I'm dropping the automated tests for now. I might find a
way to run WebExtension tests using the new Chrome and Firefox
headless modes.
- The extension should look and work the same way on all platforms.
- There is no vendor-specific code of any kind anymore 🦄